% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgpe.R
\name{fgpe}
\alias{fgpe}
\title{simple function}
\usage{
fgpe(n, p, mean, sigma, length)
}
\arguments{
\item{n}{n is the number of observations}

\item{p}{p is the dimension of graphical model}

\item{mean}{mean is a mean vector,}

\item{sigma}{sigma is a covariance matrix}

\item{length}{length is a tuning parameter of HOLP or SIS}
}
\value{
A list of
\itemize{
 \item{\code{gamma_hat}: estimator for regression coefficient matrix}
 \item{\code{p_valu}: p_value matrix of Hypothesis testing}
}
}
\description{
today,I create my first function,a very usrful function.
}
\details{
The graphical projection estimator (GPE) based on the partially penalized regression
}
