% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgpe.R
\name{tes.fgpe}
\alias{tes.fgpe}
\title{simple function}
\usage{
tes.fgpe(M, p, resultmcp)
}
\arguments{
\item{M}{M is the number of parallelism}

\item{p}{p is the dimension of graphical model}

\item{resultmcp}{resultmcp is a list of matrices generated based on the results of function parallelism}
}
\value{
A list of 
\itemize{
 \item{\code{EPs}: the true positive rates of Hypothesis testing}
 \item{ETs: the false positive rates of Hypothesis testing}
}
}
\description{
The false positive rates and the true positive rates based on the graphical projection estimator (GPE)
}
